requirements.txt
PyPDF2
pdf2image
Pillow
reportlab